+++ src/commands/expressions/update/add.ts
@@ -1,7 +1,4 @@
-import type {
-  AttributeValueNumber,
-  AttributeValueSet,
-} from "../../../types.js";
+import type { NativeNumber, NativeSet } from "../../../types.js";
 import type { AttributeNames } from "../../attributes/names.js";
 import type { AttributeValues } from "../../attributes/values.js";
 import type { AttributeOperand } from "../operands/name.js";
@@ -9,7 +6,7 @@ import type { ValueOperand } from "../operands/value.js";
 import type { IUpdateAction, UpdateAction } from "./action.js";
 import type { UpdateExpressionClauses } from "./clauses.js";
 
-type NumberOrSet = AttributeValueNumber | AttributeValueSet;
+type NumberOrSet = NativeNumber | NativeSet;
 
 // Note: the first operand *must* be an attribute name, and the second operand *must* be a value.
 export class AddAction implements IUpdateAction {