+++ test/integration/commands/write-transaction.spec.ts
@@ -1,12 +1,12 @@
-import { DynamoDBClient, GetItemCommand } from "@aws-sdk/client-dynamodb";
 import { expect } from "@infra-blocks/test";
-import type {
-  CreateTableParams,
-  WriteTransactionParams,
+import {
+  type CreateTableParams,
+  DynamoDbClient,
+  type WriteTransactionParams,
 } from "../../../src/index.js";
 import { dropAllTables } from "../fixtures.js";
 
-describe(DynamoDBClient.name, () => {
+describe(DynamoDbClient.name, () => {
   afterEach("clean up", dropAllTables());
 
   describe("writeTransaction", () => {
@@ -37,18 +37,13 @@ describe(DynamoDBClient.name, () => {
       };
       await client.writeTransaction(writeTransactionParams);
 
-      const testClient = this.createTestClient();
       await Promise.all(
         writeTransactionParams.writes.map(async (transactItem) => {
-          const response = await testClient.send(
-            new GetItemCommand({
-              TableName: createTableParams.name,
-              Key: { pk: { S: transactItem.item.pk } },
-            }),
-          );
-          expect(response.Item).to.deep.include({
-            pk: { S: transactItem.item.pk },
+          const item = await client.getItem({
+            table: createTableParams.name,
+            partitionKey: { name: "pk", value: transactItem.item.pk },
           });
+          expect(item).to.deep.include(transactItem.item);
         }),
       );
     });